# apps/api/src/modules/ai/routes.ts
import type { FastifyPluginAsync } from 'fastify';
import { z } from 'zod';
import { encryptSecret } from '@zerochack/auth';
import { AiGatewayError } from '@zerochack/ai-gateway';
import type { Environment } from '@zerochack/config';
import { database } from '@zerochack/database';
import { ApiError } from '../../errors.js';
import { authenticate, requireOwnerMfa, requirePermission } from '../auth/security.js';
import { writeAudit } from '../auth/service.js';
import type { AiService } from './service.js';

const uuid = z.string().uuid();
const positive = z.number().int().min(1).max(100_000);
const providerCreate = z.object({ name: z.string().trim().min(2).max(120), adapterKey: z.literal('openai-responses'), enabled: z.boolean().default(false), requestsPerMinute: positive.default(60), maxConcurrent: positive.max(100).default(4) });
const providerPatch = z.object({ enabled: z.boolean().optional(), requestsPerMinute: positive.optional(), maxConcurrent: positive.max(100).optional() }).refine((input) => Object.keys(input).length > 0);
const credentialCreate = z.object({ label: z.string().trim().min(2).max(120), secret: z.string().min(12).max(4096), enabled: z.boolean().default(true) });
const credentialPatch = z.object({ enabled: z.boolean().optional(), secret: z.string().min(12).max(4096).optional() }).refine((input) => Object.keys(input).length > 0);
const modelCreate = z.object({ name: z.string().trim().min(2).max(120), providerModel: z.string().trim().min(1).max(160), enabled: z.boolean().default(true), maxOutputTokens: z.number().int().min(256).max(32_768).default(4096), requestsPerMinute: positive.default(30), maxConcurrent: positive.max(100).default(2), inputCostMicrosPerM: z.number().int().min(0).max(2_000_000_000).default(0), outputCostMicrosPerM: z.number().int().min(0).max(2_000_000_000).default(0) });
const modelPatch = modelCreate.partial().refine((input) => Object.keys(input).length > 0);
const policyInput = z.object({ enabled: z.boolean(), requestsPerMinute: positive, requestsPerUserMinute: positive, maxConcurrent: positive.max(100), dailyCostLimitMicros: z.number().int().min(0).max(2_000_000_000) });
const chatInput = z.object({ prompt: z.string().trim().min(1).max(4000), idempotencyKey: z.string().uuid(), providerId: z.string().uuid().optional(), modelId: z.string().uuid().optional() });

function parse<T>(schema: z.ZodType<T>, value: unknown): T { const result = schema.safeParse(value); if (!result.success) throw new ApiError(400, 'VALIDATION_ERROR', 'Request validation failed', result.error.flatten()); return result.data; }
async function runAi<T>(operation: () => Promise<T>): Promise<T> { try { return await operation(); } catch (error) { if (error instanceof AiGatewayError) throw new ApiError(error.statusCode, error.code, error.message); throw error; } }

export const aiRoutes: FastifyPluginAsync<{ environment: Environment; ai: AiService }> = async (app, { environment, ai }) => {
  app.addHook('onSend', async (_request, reply) => { void reply.header('cache-control', 'private, no-store'); });
  app.addHook('preHandler', async (request) => authenticate(request, environment));

  app.post('/ai/chat', async (request) => {
    requirePermission(request, 'ai.use'); const input = parse(chatInput, request.body);
    const result = await runAi(() => ai.execute({ tenantId: request.tenantId!, userId: request.userId!, roles: request.roleNames ?? [], requestId: request.id, idempotencyKey: input.idempotencyKey, prompt: input.prompt, purpose: 'GENERAL_CHAT' }, { ...(input.providerId ? { providerId: input.providerId } : {}), ...(input.modelId ? { modelId: input.modelId } : {}) }));
    await writeAudit({ tenantId: request.tenantId, actorUserId: request.userId, requestId: request.id, action: result.duplicate ? 'ai.request_deduplicated' : 'ai.request_completed', resourceType: 'ai_usage', resourceId: result.usageId, ipAddress: request.ip });
    return { type: 'AI', authorLabel: 'AI Assistant', content: result.text, usageId: result.usageId, duplicate: result.duplicate };
  });

  const owner = (request: Parameters<typeof requireOwnerMfa>[0]) => { requireOwnerMfa(request); requirePermission(request, 'ai.providers.manage'); };
  app.get('/owner/ai/providers', async (request) => { owner(request); return database.aiProvider.findMany({ orderBy: { createdAt: 'asc' }, select: { id: true, name: true, adapterKey: true, enabled: true, healthStatus: true, requestsPerMinute: true, maxConcurrent: true, lastHealthCheckAt: true, createdAt: true, models: { select: { id: true, name: true, providerModel: true, enabled: true, maxOutputTokens: true, requestsPerMinute: true, maxConcurrent: true, inputCostMicrosPerM: true, outputCostMicrosPerM: true } }, credentials: { select: { id: true, label: true, enabled: true, healthStatus: true, lastUsedAt: true, keyVersion: true, createdAt: true } } } }); });
  app.post('/owner/ai/providers', async (request, reply) => { owner(request); const input = parse(providerCreate, request.body); const created = await database.aiProvider.create({ data: input }); await writeAudit({ tenantId: request.tenantId, actorUserId: request.userId, requestId: request.id, action: 'ai.provider_created', resourceType: 'ai_provider', resourceId: created.id, ipAddress: request.ip }); return reply.code(201).send(created); });
  app.patch('/owner/ai/providers/:providerId', async (request) => { owner(request); const providerId = parse(uuid, (request.params as { providerId?: string }).providerId); const input = parse(providerPatch, request.body); const data = { ...(input.enabled !== undefined ? { enabled: input.enabled } : {}), ...(input.requestsPerMinute !== undefined ? { requestsPerMinute: input.requestsPerMinute } : {}), ...(input.maxConcurrent !== undefined ? { maxConcurrent: input.maxConcurrent } : {}) }; const result = await database.aiProvider.updateMany({ where: { id: providerId }, data }); if (!result.count) throw new ApiError(404, 'AI_PROVIDER_NOT_FOUND', 'AI provider was not found'); await writeAudit({ tenantId: request.tenantId, actorUserId: request.userId, requestId: request.id, action: 'ai.provider_changed', resourceType: 'ai_provider', resourceId: providerId, ipAddress: request.ip, metadata: data }); return database.aiProvider.findUnique({ where: { id: providerId }, select: { id: true, name: true, adapterKey: true, enabled: true, healthStatus: true, requestsPerMinute: true, maxConcurrent: true } }); });
  app.post('/owner/ai/providers/:providerId/credentials', async (request, reply) => { owner(request); const providerId = parse(uuid, (request.params as { providerId?: string }).providerId); const input = parse(credentialCreate, request.body); if (!await database.aiProvider.findUnique({ where: { id: providerId }, select: { id: true } })) throw new ApiError(404, 'AI_PROVIDER_NOT_FOUND', 'AI provider was not found'); const created = await database.aiCredential.create({ data: { providerId, label: input.label, enabled: input.enabled, encryptedSecret: encryptSecret(input.secret, environment.AI_CREDENTIAL_ENCRYPTION_KEY) }, select: { id: true, providerId: true, label: true, enabled: true, healthStatus: true, keyVersion: true, createdAt: true } }); await writeAudit({ tenantId: request.tenantId, actorUserId: request.userId, requestId: request.id, action: 'ai.credential_created', resourceType: 'ai_credential', resourceId: created.id, ipAddress: request.ip, metadata: { providerId, label: input.label } }); return reply.code(201).send(created); });
  app.patch('/owner/ai/credentials/:credentialId', async (request) => { owner(request); const credentialId = parse(uuid, (request.params as { credentialId?: string }).credentialId); const input = parse(credentialPatch, request.body); const result = await database.aiCredential.updateMany({ where: { id: credentialId }, data: { ...(input.enabled !== undefined ? { enabled: input.enabled } : {}), ...(input.secret ? { encryptedSecret: encryptSecret(input.secret, environment.AI_CREDENTIAL_ENCRYPTION_KEY), keyVersion: { increment: 1 }, healthStatus: 'UNKNOWN' as const } : {}) } }); if (!result.count) throw new ApiError(404, 'AI_CREDENTIAL_NOT_FOUND', 'AI credential was not found'); await writeAudit({ tenantId: request.tenantId, actorUserId: request.userId, requestId: request.id, action: 'ai.credential_changed', resourceType: 'ai_credential', resourceId: credentialId, ipAddress: request.ip, metadata: { enabled: input.enabled, secretRotated: Boolean(input.secret) } }); return database.aiCredential.findUnique({ where: { id: credentialId }, select: { id: true, providerId: true, label: true, enabled: true, healthStatus: true, keyVersion: true, lastUsedAt: true } }); });
  app.post('/owner/ai/providers/:providerId/models', async (request, reply) => { owner(request); const providerId = parse(uuid, (request.params as { providerId?: string }).providerId); const input = parse(modelCreate, request.body); if (!await database.aiProvider.findUnique({ where: { id: providerId }, select: { id: true } })) throw new ApiError(404, 'AI_PROVIDER_NOT_FOUND', 'AI provider was not found'); const model = await database.aiModel.create({ data: { providerId, ...input } }); await writeAudit({ tenantId: request.tenantId, actorUserId: request.userId, requestId: request.id, action: 'ai.model_created', resourceType: 'ai_model', resourceId: model.id, ipAddress: request.ip, metadata: { providerId, name: model.name } }); return reply.code(201).send(model); });
  app.patch('/owner/ai/models/:modelId', async (request) => { owner(request); const modelId = parse(uuid, (request.params as { modelId?: string }).modelId); const input = parse(modelPatch, request.body); const data = Object.fromEntries(Object.entries(input).filter((entry) => entry[1] !== undefined)); const result = await database.aiModel.updateMany({ where: { id: modelId }, data }); if (!result.count) throw new ApiError(404, 'AI_MODEL_NOT_FOUND', 'AI model was not found'); await writeAudit({ tenantId: request.tenantId, actorUserId: request.userId, requestId: request.id, action: 'ai.model_changed', resourceType: 'ai_model', resourceId: modelId, ipAddress: request.ip, metadata: data }); return database.aiModel.findUnique({ where: { id: modelId } }); });
  app.put('/owner/ai/tenants/:tenantId/policy', async (request) => { owner(request); const tenantId = parse(uuid, (request.params as { tenantId?: string }).tenantId); const input = parse(policyInput, request.body); if (!await database.tenant.findUnique({ where: { id: tenantId }, select: { id: true } })) throw new ApiError(404, 'TENANT_NOT_FOUND', 'Tenant was not found'); const policy = await database.aiTenantPolicy.upsert({ where: { tenantId }, create: { tenantId, ...input }, update: input }); await writeAudit({ tenantId, actorUserId: request.userId, requestId: request.id, action: 'ai.tenant_policy_changed', resourceType: 'tenant', resourceId: tenantId, ipAddress: request.ip, metadata: { enabled: input.enabled } }); return policy; });
  app.get('/owner/ai/usage', async (request) => { requireOwnerMfa(request); requirePermission(request, 'ai.usage.read'); const query = parse(z.object({ tenantId: z.string().uuid().optional(), take: z.coerce.number().int().min(1).max(500).default(100) }), request.query); return database.aiUsage.findMany({ where: query.tenantId ? { tenantId: query.tenantId } : {}, orderBy: { createdAt: 'desc' }, take: query.take, select: { id: true, tenantId: true, userId: true, providerId: true, modelId: true, requestId: true, status: true, inputTokens: true, outputTokens: true, latencyMs: true, estimatedCostMicros: true, errorCode: true, createdAt: true, completedAt: true } }); });
};